backend/requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2